+++ Code/Keng/Graphics/include/Keng/Graphics/Resource/TextureParameters.h
@@ -9,8 +9,8 @@ namespace keng::graphics
     class TextureParameters
     {
     public:
-        uint32_t width = 0;
-        uint32_t height = 0;
+        size_t width = 0;
+        size_t height = 0;
         TextureUsage usage = TextureUsage::ShaderResource;
         FragmentFormat format = FragmentFormat::R8_G8_B8_A8_UNORM;
         CpuAccessFlags cpuAccessFlags = CpuAccessFlags::None;
